+++ src/File.cpp
@@ -9,8 +9,11 @@
 
 namespace RedFS {
 
-File::File(std::filesystem::path p_path, std::filesystem::path p_absolute_path)
-    : path(std::move(p_path)), absolute_path(std::move(p_absolute_path)) {}
+File::File(SharedMutex p_mutex, std::filesystem::path p_path,
+           std::filesystem::path p_absolute_path)
+    : mutex(std::move(p_mutex)),
+      path(std::move(p_path)),
+      absolute_path(std::move(p_absolute_path)) {}
 
 Red::CString File::get_path() const {
   return path.string();
@@ -35,24 +38,29 @@ uint64_t File::get_size() const {
 }
 
 Red::CString File::read_as_text() {
+  mutex->lock();
   std::ifstream stream;
 
   stream.open(absolute_path);
   if (!stream.is_open()) {
+    mutex->unlock();
     return {};
   }
   std::stringstream data;
 
   data << stream.rdbuf();
   stream.close();
+  mutex->unlock();
   return data.str();
 }
 
 Red::DynArray<Red::CString> File::read_as_lines() {
+  mutex->lock();
   std::ifstream stream;
 
   stream.open(absolute_path);
   if (!stream.is_open()) {
+    mutex->unlock();
     return {};
   }
   Red::DynArray<Red::CString> lines;
@@ -60,11 +68,14 @@ Red::DynArray<Red::CString> File::read_as_lines() {
 
   while (std::getline(stream, line)) {
     if (stream.fail() || stream.bad()) {
-      break;
+      stream.close();
+      mutex->unlock();
+      return lines;
     }
     lines.EmplaceBack(line);
   }
   stream.close();
+  mutex->unlock();
   return lines;
 }
 
@@ -80,25 +91,30 @@ Red::Handle<Red::IScriptable> File::read_as_json() {
 
 bool File::write_text(const Red::CString& p_text,
                       const Red::Optional<FileSystemWriteMode>& p_mode) {
+  mutex->lock();
   std::ios_base::openmode mode = get_mode(p_mode.value);
   std::ofstream stream;
 
   stream.open(absolute_path, mode);
   if (!stream.is_open()) {
+    mutex->unlock();
     return false;
   }
   stream << p_text.c_str();
   stream.close();
+  mutex->unlock();
   return true;
 }
 
 bool File::write_lines(const Red::DynArray<Red::CString>& p_lines,
                        const Red::Optional<FileSystemWriteMode>& p_mode) {
+  mutex->lock();
   std::ios_base::openmode mode = get_mode(p_mode.value);
   std::ofstream stream;
 
   stream.open(absolute_path, mode);
   if (!stream.is_open()) {
+    mutex->unlock();
     return false;
   }
   for (uint32_t i = 0; i < p_lines.size; i++) {
@@ -108,6 +124,7 @@ bool File::write_lines(const Red::DynArray<Red::CString>& p_lines,
     }
   }
   stream.close();
+  mutex->unlock();
   return true;
 }