+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+------------------------
+
+## [0.8.0] - 2024-05-17
 ### Added
 - asynchronous read/write operations with `AsyncFile`.
 - callback handling with `FilePromise` when operation passes/fails.
@@ -12,10 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   It can be accessed with `FileSystem.GetSharedStorage()` without restrictions.
 
 ### Changed
-- migrate storages from `<game>/red4ext/plugins/RedFileSystem/storages/` to 
+- migrate storages from `<game>/red4ext/plugins/RedFileSystem/storages/` to
   `<game>/r6/storages/` at startup.
 - read/write operations with `File` are now thread safe.
-- the name `shared` when using `FileSystem.GetStorage()` is now reserved and 
+- the name `shared` when using `FileSystem.GetStorage()` is now reserved and
   forbidden.
 
 ### Fixed
@@ -89,7 +93,8 @@ This version is not compatible with v0.2.0 and below, see the README and
 - read / write / stringify Json data.
 
 <!-- Table of releases -->
-[Unreleased]: https://github.com/rayshader/cp2077-red-filesystem/compare/v0.7.1...HEAD
+[Unreleased]: https://github.com/rayshader/cp2077-red-filesystem/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/rayshader/cp2077-red-filesystem/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/rayshader/cp2077-red-filesystem/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/rayshader/cp2077-red-filesystem/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/rayshader/cp2077-red-filesystem/compare/v0.5.0...v0.6.0